Make it possible to build with cmake
Summary:
Introduces a CMake-based build system that is compatible with
binary-factory.kde.org and works like the rest of KDE applications.
It doesn't make use of the taglib here but fetches the sources on build
time on android. For other platforms it will look for the dependency in
the system like the rest of depedencencies.

Test Plan: Can build both using cmake and qmake on linux and android.

Reviewers: camiloh

Reviewed By: camiloh

Differential Revision: https://phabricator.kde.org/D12286